CancellatioinToken and ApplicationManager moved
