+++ src/core/exchanges/general/order/create_websocket_based.rs
@@ -3,12 +3,12 @@ use log::{error, info};
 use tokio::sync::oneshot;
 
 use crate::core::{
-    exchanges::cancellation_token::CancellationToken,
     exchanges::common::ExchangeError,
     exchanges::common::ExchangeErrorType,
     exchanges::common::RestRequestOutcome,
     exchanges::general::exchange::Exchange,
     exchanges::general::exchange::RequestResult,
+    lifecycle::cancellation_token::CancellationToken,
     orders::order::ClientOrderId,
     orders::order::ExchangeOrderId,
     orders::{fill::EventSourceType, order::OrderCreating},
