+++ src/core/lifecycle/application_manager.rs
