+++ src/core/lifecycle/launcher.rs
@@ -6,8 +6,6 @@ use futures::{future::join_all, FutureExt};
 use log::{error, info};
 use tokio::sync::{broadcast, oneshot};
 
-use crate::core::exchanges::application_manager::ApplicationManager;
-use crate::core::exchanges::cancellation_token::CancellationToken;
 use crate::core::exchanges::common::{CurrencyPair, ExchangeAccountId, ExchangeId};
 use crate::core::exchanges::events::{ExchangeEvent, ExchangeEvents, CHANNEL_MAX_EVENTS_COUNT};
 use crate::core::exchanges::general::exchange_creation::create_exchange;
@@ -17,6 +15,8 @@ use crate::core::exchanges::general::{
 use crate::core::exchanges::timeouts::timeout_manager::TimeoutManager;
 use crate::core::exchanges::traits::ExchangeClientBuilder;
 use crate::core::internal_events_loop::InternalEventsLoop;
+use crate::core::lifecycle::application_manager::ApplicationManager;
+use crate::core::lifecycle::cancellation_token::CancellationToken;
 use crate::core::lifecycle::trading_engine::{EngineContext, TradingEngine};
 use crate::core::logger::init_logger;
 use crate::core::order_book::local_snapshot_service::LocalSnapshotsService;
