+++ src/core/lifecycle/mod.rs
@@ -1,3 +1,5 @@
+pub mod application_manager;
+pub mod cancellation_token;
 pub mod launcher;
 pub mod shutdown;
 pub mod trading_engine;
