+++ tests/binance/cancel_order.rs
@@ -1,14 +1,13 @@
 use std::collections::HashMap;
 
 use chrono::Utc;
-use mmb_lib::core::exchanges::application_manager::ApplicationManager;
+use mmb_lib::core::exchanges::events::AllowedEventSourceType;
 use mmb_lib::core::exchanges::general::exchange::*;
 use mmb_lib::core::exchanges::general::features::*;
 use mmb_lib::core::exchanges::{binance::binance::*, general::commission::Commission};
-use mmb_lib::core::exchanges::{
-    cancellation_token::CancellationToken, events::AllowedEventSourceType,
-};
 use mmb_lib::core::exchanges::{common::*, timeouts::timeout_manager::TimeoutManager};
+use mmb_lib::core::lifecycle::application_manager::ApplicationManager;
+use mmb_lib::core::lifecycle::cancellation_token::CancellationToken;
 use mmb_lib::core::logger::init_logger;
 use mmb_lib::core::orders::order::*;
 use mmb_lib::core::settings;
