+++ tests/binance/wait_cancel_order.rs
@@ -1,13 +1,12 @@
 use crate::get_binance_credentials_or_exit;
 use chrono::Utc;
-use mmb_lib::core::exchanges::application_manager::ApplicationManager;
 use mmb_lib::core::exchanges::common::*;
+use mmb_lib::core::exchanges::events::AllowedEventSourceType;
 use mmb_lib::core::exchanges::general::exchange::*;
 use mmb_lib::core::exchanges::general::features::*;
 use mmb_lib::core::exchanges::{binance::binance::*, general::commission::Commission};
-use mmb_lib::core::exchanges::{
-    cancellation_token::CancellationToken, events::AllowedEventSourceType,
-};
+use mmb_lib::core::lifecycle::application_manager::ApplicationManager;
+use mmb_lib::core::lifecycle::cancellation_token::CancellationToken;
 use mmb_lib::core::logger::init_logger;
 use mmb_lib::core::orders::order::*;
 use mmb_lib::core::settings;
